refactor: Home.js is slimmed down into componnets
