+++ client/components/Home.js
@@ -1,11 +1,12 @@
 import React, { Component, useState, useEffect } from 'react';
 import { Link, Redirect } from 'react-router-dom';
-import Particles from 'react-particles-js';
 import { useAuthState } from 'react-firebase-hooks/auth';
 
 import Loading from './Loading';
 import SignIn from './SignIn';
 import SignOut from './SignOut';
+import BackgroundParticles from './Particles.js';
+import InstructionsToggle from './Instructions';
 
 import firebase from 'firebase/app';
 import 'firebase/firestore';
@@ -23,7 +24,7 @@ const Home = (props) => {
 		console.log('user-->', user);
 		console.log('props-->', props);
 		if (user) props.setNewUser(user);
-	});
+	}, [user]);
 	const [enableOutline, setEnableOutline] = useState(false);
 	const [showInstructions, setShowInstructions] = useState(false);
 	const [redirectTo, setRedirectTo] = useState('');
@@ -42,57 +43,12 @@ const Home = (props) => {
 		setShowInstructions(!showInstructions);
 	};
 
-	// JSX
+	// Background Animation,
+	// Conditional Auth render
 	return (
 		<>
 			<div style={{ textAlign: 'center', marginTop: '15%', fontSize: '60px' }}>
-				<Particles
-					id='particles-js'
-					params={{
-						particles: {
-							line_linked: {
-								shadow: {
-									enable: true,
-									color: `${randomColor}`,
-									blur: 1.5
-								}
-							},
-							interactivity: {
-								detect_on: 'canvas',
-								events: {
-									onhover: {
-										enable: true,
-										mode: 'grab'
-									},
-									resize: true
-								},
-								modes: {
-									grab: {
-										distance: 400,
-										line_linked: {
-											opacity: 1
-										}
-									}
-								}
-							},
-							move: {
-								enable: true,
-								speed: 0.6,
-								direction: 'bottom-left',
-								random: false,
-								straight: false,
-								out_mode: 'bounce',
-								bounce: false,
-								attract: {
-									enable: false,
-									rotateX: 600,
-									rotateY: 1200
-								}
-							},
-							retina_detect: true
-						}
-					}}
-				/>
+				<BackgroundParticles />
 				<header style={{ textShadow: '2px 6px 6 rgba(218, 217, 217, 0.346)' }}>
 					a e t h e r
 				</header>
@@ -100,7 +56,7 @@ const Home = (props) => {
 
 				{loading ? (
 					<Loading />
-				) : props.user.displayName ? (
+				) : auth.currentUser !== null ? (
 					<>
 						<Link to='/sesh'>
 							<button
@@ -125,47 +81,11 @@ const Home = (props) => {
 						<br />
 						<SignOut user={user} enableOutline={enableOutline} />
 						<br />
-						{showInstructions ? (
-							<div
-								style={{
-									textAlign: 'center',
-									fontSize: '15px',
-									fontFamily: 'Source Code Pro, monospace'
-								}}
-							>
-								<button
-									type='button'
-									className={
-										enableOutline ? 'home-btn' : 'no-outline-on-focus home-btn'
-									}
-									onClick={showDirections}
-									style={{ textAlign: 'center', marginTop: '1%', fontSize: '12px' }}
-								>
-									close instructions
-								</button>
-								<h4>“One band. One Sound.”</h4>
-								<p>
-									You have the unique opportunity to collaborate in realtime with other
-									musicians
-								</p>
-								<p>Click a sound object Once to hear its sound.</p>
-								<p>Click and Drag a sound object onto the Jam-Space</p>
-								<p>When the Hammer strikes it, hear the sounds!</p>
-								<p>
-									Exlore the music you can create with your friends or match with users
-									from around the world.
-								</p>
-							</div>
-						) : (
-							<button
-								type='button'
-								className={enableOutline ? 'home-btn' : 'no-outline-on-focus home-btn'}
-								onClick={showDirections}
-								style={{ textAlign: 'center', marginTop: '1%', fontSize: '12px' }}
-							>
-								view instructions
-							</button>
-						)}
+						<InstructionsToggle
+							enableOutline={enableOutline}
+							showInstructions={showInstructions}
+							showDirections={showDirections}
+						/>
 					</>
 				) : (
 					<>
@@ -181,47 +101,11 @@ const Home = (props) => {
 						<br />
 						<SignIn enableOutline={enableOutline} />
 						<br />
-						{showInstructions ? (
-							<div
-								style={{
-									textAlign: 'center',
-									fontSize: '15px',
-									fontFamily: 'Source Code Pro, monospace'
-								}}
-							>
-								<button
-									type='button'
-									className={
-										enableOutline ? 'home-btn' : 'no-outline-on-focus home-btn'
-									}
-									onClick={showDirections}
-									style={{ textAlign: 'center', marginTop: '1%', fontSize: '12px' }}
-								>
-									close instructions
-								</button>
-								<h4>“One band. One Sound.”</h4>
-								<p>
-									You have the unique opportunity to collaborate in realtime with other
-									musicians
-								</p>
-								<p>Click a sound object Once to hear its sound.</p>
-								<p>Click and Drag a sound object onto the Jam-Space</p>
-								<p>When the Hammer strikes it, hear the sounds!</p>
-								<p>
-									Exlore the music you can create with your friends or match with users
-									from around the world.
-								</p>
-							</div>
-						) : (
-							<button
-								type='button'
-								className={enableOutline ? 'home-btn' : 'no-outline-on-focus home-btn'}
-								onClick={showDirections}
-								style={{ textAlign: 'center', marginTop: '1%', fontSize: '12px' }}
-							>
-								view instructions
-							</button>
-						)}
+						<InstructionsToggle
+							enableOutline={enableOutline}
+							showInstructions={showInstructions}
+							showDirections={showDirections}
+						/>
 					</>
 				)}
 				<br />
@@ -229,10 +113,6 @@ const Home = (props) => {
 		</>
 	);
 };
-
-
-// Helper Function
-const randomColor = Math.floor(Math.random() * 16777215).toString(16);
 
 // Connect Redux
 const mapStateToProps = (state) => ({ user: state.user });
