+++ client/components/SignIn.js
@@ -8,7 +8,7 @@ const SignIn = (props) => {
 		const provider = new firebase.auth.GoogleAuthProvider();
 		//instatiate new auth token
 		auth.signInWithRedirect(provider);
-		//prompts redirect to google login, then back
+		//prompts redirect to and back from google login
 	};
 
 	return (
